Release v2.1.0: Add decimal row rate support and UI improvements
- Add decimal support for row rate parameter (e.g., 2400.5 kHz)
- Expand row rate range from 1-6144 to 1-48000 kHz
- Create ValidatedFloat descriptor for float validation
- Update CSV parser to handle float row rate values
- Enhance UI validator to accept decimal input
- Make interface parameter entry boxes 20% wider for better readability
- Add custom UI label 'Row Rate [kHz] (1-48000)'
- Add scratch/ directory to .gitignore

+++ src/models/bus_model.py
@@ -91,7 +91,7 @@ class BusModel:
     """
     num_rows: int
     num_columns: int
-    row_rate: int = 3072  # kHz, default value
+    row_rate: float = 3072.0  # kHz, default value
     bits: List[BitInfo] = field(default_factory=list)
 
     # Clash tracking
